Added YAML configuration file
The configuration file specifies the gesture dataset in YAML. You can customize the dataset by adding or deleting gestures in the config file. Used a library to parse YAML files and stored the result in a Gesture object. Used the information gathered from the config file to populate and display gesture info to the user in the application.